[checkrunner] allow namespace items to 'force' override existing namespace entries.
